refactor: 날씨 관련 로직을 ScheduleService에서 분리
